feat(CardSlid): change BrandProps to GlobalTypes affects: @naturacosmeticos/natds-react-js DSY-3194
