+++ packages/react/src/Templates/GlobalTypes/index.ts
@@ -1,0 +1,10 @@
+export type BrandProps =
+    'aesop'
+    | 'avon'
+    | 'avon_v2'
+    | 'biome'
+    | 'natura'
+    | 'theBodyShop'
+    | 'consultoriaDeBeleza'
+    | 'natura_v2'
+    | 'casaEestilo'
